lib,src: refactor src/node.js into internal files
PR-URL: https://github.com/nodejs/node/pull/5103

Conflicts:
	node.gyp
	src/node.js
