+++ node.gyp
@@ -76,6 +76,10 @@
       'lib/internal/linkedlist.js',
       'lib/internal/net.js',
       'lib/internal/module.js',
+      'lib/internal/process/next_tick.js',
+      'lib/internal/process/promises.js',
+      'lib/internal/process/stdio.js',
+      'lib/internal/process.js',
       'lib/internal/repl.js',
       'lib/internal/socket_list.js',
       'lib/internal/util.js',
